requirements.txt
flask==3.0.2
weaviate-client==4.14.4
openai==1.12.0
opik==0.1.0
requests>=2.31.0
python-dotenv>=1.0.0 